Refactor: rename template classes and package -remove template word from the class names -rename templates package to abstracts
