+++ lib/src/main/kotlin/trees/abstracts/BalanceBSTree.kt
@@ -1,0 +1,4 @@
+package trees.abstracts
+
+abstract class BalanceBSTree<K : Comparable<K>, V, Vertex_t : Vertex<K, V, Vertex_t>> :
+    BSTree<K, V, Vertex_t>()
